pyproject.toml
[project]
name = "chuqin"
version = "0.0.1"
readme = "README.md"

requires-python = ">=3.11"

dependencies = [
    "pymupdf>=1.24.0",
    "python-pptx>=1.0.2",
    "typer>=0.16.0",
    "pywebview>=6.0.0",
]

[project.optional-dependencies]
build = [
    "pyinstaller>=6.0.0",
]
dev = [
    "mypy",
    "ruff",
    "types-pyinstaller",
]
test = [
    "pytest",
]

[project.scripts]
chuqin = "app.cli:main"

[project.urls]
Homepage = "https://github.com/hezhangjian/ChuQin"
Repository = "https://github.com/hezhangjian/ChuQin"
Documentation = "https://github.com/hezhangjian/ChuQin#readme"

[tool.mypy]
python_version = "3.11"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "launcher*"]
exclude = ["portal*"]

[tool.ruff]
line-length = 120
